+++ RELEASE_NOTES.md
@@ -20,6 +20,7 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
 - Added instance number setting to PHD2 to be able to have N.I.N.A. autostart PHD2 with multiple instances and not just the first one
 - Changed Message Box Instruction to include "Stop Sequence" button along with "Continue" to be able to stop the sequence in case this is used as a checkpoint
 - When PHD2 Configuration Changes happen, N.I.N.A. will now requery the pixel scale in case the focal length or pixel size setting of the guider has changed
+- Improved layout of plate solve pop up during centering and added a status indicator inside the window as well
 
 ## Features
 - Enhanced Meridian Flip options with a toggle to automatically rotate the target in the imaging tab for displaying purposes